refactor: extract CacheStore and decompose LutamlEaXmiBase
Add thread-safe CacheStore with Mutex and FIFO eviction, replacing
class-variable caches. Decompose the 600-line LutamlEaXmiBase god module
into five focused modules: XmiCache, XmiConfig, XmiPackageFilter,
XmiContextBuilder, XmiRenderer. Add collection_for helper to Config::Package.
