+++ lib/metanorma/plugin/lutaml/lutaml_ea_xmi_base.rb
@@ -1,31 +1,19 @@
 # frozen_string_literal: true
 
-require "liquid"
 require "asciidoctor"
 require "asciidoctor/reader"
-require "lutaml"
-require "lutaml/uml"
-require "lutaml/formatter"
-require_relative "utils"
 
 module Metanorma
   module Plugin
     module Lutaml
       module LutamlEaXmiBase
         include Utils
+        include XmiCache
+        include XmiConfig
+        include XmiPackageFilter
+        include XmiContextBuilder
+        include XmiRenderer
 
-        LIQUID_INCLUDE_PATH = File.join(
-          Gem.loaded_specs["metanorma-plugin-lutaml"].full_gem_path,
-          "lib", "metanorma", "plugin", "lutaml", "liquid_templates"
-        )
-        DEFAULT_RENDER_INCLUDE = "packages"
-        RENDER_STYLES_INCLUDES = {
-          "default" => "packages",
-          "entity_list" => "packages_entity_list",
-          "entity_list_class" => "packages_entity_list_class",
-          "data_dictionary" => "packages_data_dictionary",
-        }.freeze
-        RENDER_STYLE_ATTRIBUTE = "render_style"
         SUPPORTED_NESTED_MACRO = %w[
           before diagram_include_block after include_block package_text
         ].freeze
@@ -42,10 +30,6 @@ module Metanorma
           $                    # End of the pattern
         }x
 
-        # search document for block `lutaml_ea_xmi`
-        # or `lutaml_uml_datamodel_description`
-        # read include directives that goes after that in block and transform
-        # into yaml2text blocks
         def process(document, reader)
           r = Asciidoctor::PreprocessorNoIfdefsReader.new document, reader.lines
           input_lines = r.readlines.to_enum
@@ -54,55 +38,6 @@ module Metanorma
         end
 
         private
-
-        def lutaml_document_from_file_or_cache(document, file_path, yaml_config, yaml_config_path = nil) # rubocop:disable Metrics/AbcSize,Metrics/MethodLength,Layout/LineLength
-          full_path = Utils.relative_file_path(document, file_path)
-          if document.attributes["lutaml_xmi_cache"] &&
-              document.attributes["lutaml_xmi_cache"][full_path]
-            return document.attributes["lutaml_xmi_cache"][full_path]
-          end
-
-          yaml_config.ea_extension&.each do |ea_extension_path|
-            ea_extension_full_path = File.expand_path(
-              ea_extension_path, File.dirname(yaml_config_path)
-            )
-            unless Xmi::EaRoot.loaded_extensions.value?(ea_extension_full_path)
-              Xmi::EaRoot.load_extension(ea_extension_full_path)
-            end
-          end
-
-          guidance = get_guidance(document, yaml_config.guidance)
-          cache_key = [full_path, guidance]
-
-          @@lutaml_doc_cache ||= {}
-          result_document = @@lutaml_doc_cache[cache_key] ||= parse_result_document(
-            full_path, guidance
-          )
-
-          document.attributes["lutaml_xmi_cache"] ||= {}
-          document.attributes["lutaml_xmi_cache"][full_path] = result_document
-          result_document
-        end
-
-        def get_guidance(document, guidance_config)
-          return unless guidance_config
-
-          guidance_yaml = Utils.relative_file_path(document, guidance_config)
-          guidance = Metanorma::Plugin::Lutaml::Config::Guidance.from_yaml(
-            File.read(guidance_yaml, encoding: "UTF-8"),
-          )
-          guidance.to_yaml_hash
-        end
-
-        def parse_yaml_config_file(document, file_path)
-          return Metanorma::Plugin::Lutaml::Config::Root.new if file_path.nil?
-
-          relative_file_path = Utils.relative_file_path(document, file_path)
-
-          Metanorma::Plugin::Lutaml::Config::Root.from_yaml(
-            File.read(relative_file_path, encoding: "UTF-8"),
-          )
-        end
 
         def get_xmi_path(parent, target, attrs)
           return get_path_from_index(parent, attrs["index"]) if attrs["index"]
@@ -175,7 +110,6 @@ module Metanorma
           index = xmi_or_index.match(/index=(.*)/)
 
           if index
-            # load lutaml index
             index_name = index[1]
 
             if document.attributes["lutaml_xmi_index"][index_name].nil? ||
@@ -201,53 +135,6 @@ module Metanorma
             Utils.relative_file_path(document, config_yaml_path),
           )
           [lutaml_document, yaml_config]
-        end
-
-        def fill_in_entities_refs_attributes(document, # rubocop:disable Metrics/AbcSize,Metrics/CyclomaticComplexity,Metrics/MethodLength
-          lutaml_document, _options)
-          # render_style = options.fetch(RENDER_STYLE_ATTRIBUTE, "default")
-          all_children_packages = lutaml_document.packages
-            .map(&:children_packages).flatten
-          package_flat_packages = lambda do |pks|
-            pks.to_h do |package|
-              [package.name, package.xmi_id]
-            end
-          end
-          children_pks = package_flat_packages.call(all_children_packages)
-          ref_dictionary = package_flat_packages.call(lutaml_document.packages)
-            .merge(children_pks)
-          %w[class enum data_type].each do |type|
-            package_flat = lambda do |pks|
-              pks.each_with_object({}) do |package, res|
-                plural = type == "class" ? "classes" : "#{type}s"
-                package.send(plural).map do |entity|
-                  res["#{type}:#{package.name}:#{entity.name}"] = entity.xmi_id
-                end
-              end
-            end
-            children_pks_diags = package_flat.call(all_children_packages)
-            ref_dictionary = ref_dictionary
-              .merge(package_flat.call(lutaml_document.packages)
-                                      .merge(children_pks_diags))
-          end
-          document.attributes["lutaml_entity_id"] = ref_dictionary
-        end
-
-        def fill_in_diagrams_attributes(document, lutaml_document) # rubocop:disable Metrics/AbcSize,Metrics/MethodLength
-          package_flat_diagrams = lambda do |pks|
-            pks.each_with_object({}) do |package, res|
-              package.diagrams.map do |diag|
-                res["#{package.name}:#{diag.name}"] = diag.xmi_id
-              end
-            end
-          end
-          children_pks_diags = package_flat_diagrams.call(
-            lutaml_document.packages.map(&:children_packages).flatten,
-          )
-
-          document.attributes["lutaml_figure_id"] = package_flat_diagrams
-            .call(lutaml_document.packages)
-            .merge(children_pks_diags)
         end
 
         def collect_block_lines(input_lines, end_mark)
@@ -293,305 +180,6 @@ module Metanorma
                       "type" => name, "package" => package }.merge(attrs))
           end
           additional_context
-        end
-
-        def package_level(lutaml_document, level)
-          return lutaml_document if level <= 0
-
-          package_level(lutaml_document.packages.first, level - 1)
-        end
-
-        def create_context_object( # rubocop:disable Metrics/AbcSize,Metrics/MethodLength
-          lutaml_document, additional_context, options,
-          context_name = "context"
-        )
-          root_package = package_level(lutaml_document.to_liquid,
-                                       options.package_root_level || 1)
-          contexts = create_default_context_object(
-            context_name, root_package, additional_context, options
-          )
-
-          if options.packages.nil?
-            contexts[context_name]["render_nested_packages"] = true
-            contexts[context_name]["packages"] = root_package.packages
-
-            return contexts
-          end
-
-          all_packages = [root_package, *root_package.children_packages]
-          contexts[context_name].merge!(
-            {
-              "packages" => sort_and_filter_out_packages(all_packages, options),
-              "package_entities" => package_hash(options, "render_entities"),
-              "package_skip_sections" => package_hash(options, "skip_tables"),
-              "render_nested_packages" => !!options.render_nested_packages,
-            },
-          )
-
-          contexts
-        end
-
-        def create_default_context_object(
-          context_name, root_package, additional_context, options
-        )
-          contexts = {}
-          contexts[context_name] = {
-            "name" => root_package.name,
-            "root_packages" => [root_package],
-            "additional_context" => additional_context
-              .merge("external_classes" => options.external_classes),
-            "skip_unrecognized_connector" => !!options
-              .skip_unrecognized_connector,
-          }
-          contexts
-        end
-
-        def package_hash(options, key)
-          return {} unless options.packages
-
-          result = {}
-          packages = options.packages.reject { |p| p.send(key.to_sym).nil? }
-          packages.each do |p|
-            result[p.name] = p.send(key.to_sym).to_h { |n| [n, true] }
-          end
-          result
-        end
-
-        def sort_and_filter_out_packages(all_packages, options) # rubocop:disable Metrics/AbcSize,Metrics/CyclomaticComplexity,Metrics/MethodLength,Metrics/PerceivedComplexity
-          return all_packages if options.packages.nil?
-
-          result = []
-          # Step one - filter out all skipped packages
-          all_packages =
-            filter_out_all_skipped_packages(options, all_packages)
-
-          # Step two - select supplied packages by pattern
-          select_supplied_packages_by_pattern(options, all_packages,
-                                              result)
-        end
-
-        def filter_out_all_skipped_packages(options, all_packages) # rubocop:disable Metrics/AbcSize
-          return all_packages if options.skip.nil?
-
-          options.skip.each do |skip_package|
-            entity_regexp = config_entity_regexp(skip_package)
-            all_packages.delete_if do |package|
-              package.name =~ entity_regexp
-            end
-          end
-
-          all_packages
-        end
-
-        def select_supplied_packages_by_pattern(options, all_packages, result) # rubocop:disable Metrics/AbcSize,Metrics/MethodLength
-          options.packages.each do |package|
-            entity_regexp = config_entity_regexp(package.name)
-            all_packages.each do |p|
-              if p.name&.match?(entity_regexp)
-                result.push(p)
-                all_packages.delete_if do |nest_package|
-                  nest_package.name == p.name
-                end
-              end
-            end
-          end
-
-          result
-        end
-
-        def config_entity_regexp(entity)
-          additional_sym = ".*" if /\*$/.match?(entity)
-          %r{^#{Regexp.escape(entity.delete('*'))}#{additional_sym}$}
-        end
-
-        def model_representation(lutaml_doc, document, add_context, options) # rubocop:disable Metrics/MethodLength
-          fill_in_entities_refs_attributes(document, lutaml_doc, options)
-          render_result, errors = Utils.render_liquid_string(
-            template_string: template(options.section_depth || 2,
-                                      options.render_style,
-                                      options.include_root),
-            contexts: create_context_object(lutaml_doc,
-                                            add_context,
-                                            options,
-                                            "context"),
-            document: document,
-            include_path: template_path(document, options.template_path),
-          )
-          Utils.notify_render_errors(document, errors)
-          render_result.split("\n")
-        end
-
-        def template_path(document, template_path)
-          return LIQUID_INCLUDE_PATH if template_path.nil?
-
-          Utils.relative_file_path(document, template_path)
-        end
-
-        def template(section_depth, render_style, include_root)
-          include_name = RENDER_STYLES_INCLUDES.fetch(render_style,
-                                                      DEFAULT_RENDER_INCLUDE)
-          result = ""
-          if include_root
-            result += <<~LIQUID
-              {% include "#{include_name}", package_skip_sections: context.package_skip_sections, package_entities: context.package_entities, context: context.root_packages, additional_context: context.additional_context, render_nested_packages: false %}
-            LIQUID
-          end
-          result + <<~LIQUID
-            {% include "#{include_name}", depth: #{section_depth}, package_skip_sections: context.package_skip_sections, package_entities: context.package_entities, context: context, additional_context: context.additional_context, render_nested_packages: context.render_nested_packages %}
-          LIQUID
-        end
-
-        def render_table(context, context_name, parent, attrs) # rubocop:disable Metrics/AbcSize,Metrics/MethodLength
-          table_tmpl = get_template(parent.document, attrs)
-          table_tmpl.assigns[context_name] = context
-
-          if attrs["external_data"]
-            data_array = attrs["external_data"].split(";")
-            data_array.each do |data_item|
-              context_name, external_data_path = data_item.split(":")
-              external_data = content_from_file(
-                parent.document, external_data_path.strip
-              )
-              table_tmpl.assigns[context_name.strip] = external_data
-            end
-          end
-
-          rendered_table = table_tmpl.render
-          block = create_open_block(parent, "", attrs)
-          parse_content(block, rendered_table, attrs)
-        end
-
-        def get_template(document, attrs)
-          template = get_default_template
-          template = attrs["template"] if attrs["template"]
-
-          rel_tmpl_path = Utils.relative_file_path(
-            document, template
-          )
-
-          ::Liquid::Template.parse(File.read(rel_tmpl_path))
-        end
-
-        def get_name_path(attrs)
-          return attrs["path"] if attrs["path"]
-
-          return "#{attrs['package']}::#{attrs['name']}" if attrs["package"]
-
-          attrs["name"]
-        end
-
-        # The class methods `serialize_generalization_by_name` and
-        # `serialize_enumeration_by_name` were removed from
-        # `Lutaml::Xmi::Parsers::Xml` in lutaml 0.10. The replacements below
-        # rebuild the same shape of result by reusing the still-available
-        # path-aware finders on the parser instance and the new
-        # `XmiLookupService`, then bridging to the UML object that the
-        # current `KlassDrop` / `EnumDrop` constructors expect.
-        def serialize_klass_drop_by_name(xmi_path, name, document = nil,
-guidance = nil)
-          parser, uml_doc = build_uml_document(xmi_path, document)
-          root_model_name = parser.xmi_root_model.model.name
-          raw_klass = find_packaged_klass(parser.xmi_index, name,
-                                          root_model_name: root_model_name)
-          warn "Class not found for name: #{name}" if raw_klass.nil?
-          klass = raw_klass && find_uml_node_by_xmi_id(
-            uml_doc, raw_klass.id, :classes
-          )
-          ::Lutaml::Xmi::LiquidDrops::KlassDrop.new(
-            klass, guidance, build_drop_options(parser)
-          )
-        end
-
-        def serialize_enum_drop_by_name(xmi_path, name, document = nil)
-          parser, uml_doc = build_uml_document(xmi_path, document)
-          raw_enum = find_packaged_enum(parser.xmi_index, name)
-          warn "Enumeration not found for name: #{name}" if raw_enum.nil?
-          enum = raw_enum && find_uml_node_by_xmi_id(
-            uml_doc, raw_enum.id, :enums
-          )
-          ::Lutaml::Xmi::LiquidDrops::EnumDrop.new(
-            enum, build_drop_options(parser)
-          )
-        end
-
-        def build_uml_document(xmi_path, _document = nil)
-          @@uml_doc_cache ||= {}
-          if @@uml_doc_cache[xmi_path]
-            return @@uml_doc_cache[xmi_path]
-          end
-
-          xmi_model = ::Xmi::Sparx::Root.parse_xml(File.read(xmi_path))
-          parser = ::Lutaml::Xmi::Parsers::Xml.new
-          result = [parser, parser.parse(xmi_model)]
-          @@uml_doc_cache[xmi_path] = result
-          result
-        end
-
-        def build_drop_options(parser)
-          lookup = ::Lutaml::Xmi::XmiLookupService.new(
-            parser.xmi_root_model, parser.id_name_mapping
-          )
-          {
-            xmi_root_model: parser.xmi_root_model,
-            id_name_mapping: parser.id_name_mapping,
-            lookup: lookup,
-            with_gen: true,
-            with_absolute_path: true,
-          }
-        end
-
-        def find_uml_node_by_xmi_id(container, xmi_id, collection)
-          found = container.public_send(collection)
-            .find { |node| node.xmi_id == xmi_id }
-          return found if found
-
-          container.packages.each do |pkg|
-            nested = find_uml_node_by_xmi_id(pkg, xmi_id, collection)
-            return nested if nested
-          end
-          nil
-        end
-
-        def find_packaged_klass(index, path, root_model_name: nil)
-          segments = path.split("::").reject(&:empty?)
-          if root_model_name && segments.first == root_model_name
-            segments.shift
-          end
-          if segments.one?
-            index.find_packaged_by_name_and_types(
-              segments.first, ["uml:Class", "uml:AssociationClass"]
-            )
-          else
-            find_packaged_klass_by_path(index, segments)
-          end
-        end
-
-        def find_packaged_klass_by_path(index, segments)
-          klass_name = segments.pop
-
-          candidates = ["uml:Class", "uml:AssociationClass"]
-            .flat_map { |t| index.packaged_elements_of_type(t) }
-            .select { |e| e.name == klass_name }
-
-          candidates.find do |klass|
-            match_parent_chain?(index, klass, segments)
-          end
-        end
-
-        def match_parent_chain?(index, element, parent_segments)
-          current = element
-          parent_segments.reverse_each do |pkg_name|
-            parent = index.find_parent(current.id)
-            return false unless parent && parent.name == pkg_name
-
-            current = parent
-          end
-          true
-        end
-
-        def find_packaged_enum(index, name)
-          index.packaged_elements_of_type("uml:Enumeration")
-            .find { |e| e.name == name }
         end
       end
     end
